only expose action from optional plugin xml
