+++ apps/intellij/src/main/kotlin/dev/nx/console/cloud/CIPENotificationService.kt
@@ -1,16 +1,11 @@
 package dev.nx.console.cloud
 
 import com.intellij.ide.BrowserUtil
-import com.intellij.ml.llm.core.chat.session.*
-import com.intellij.ml.llm.core.chat.ui.AIAssistantUIUtil
-import com.intellij.ml.llm.privacy.trustedStringBuilders.privacyConst
 import com.intellij.notification.*
-import com.intellij.openapi.actionSystem.AnAction
+import com.intellij.openapi.actionSystem.ActionManager
 import com.intellij.openapi.actionSystem.AnActionEvent
-import com.intellij.openapi.application.EDT
 import com.intellij.openapi.components.Service
 import com.intellij.openapi.components.service
-import com.intellij.openapi.diagnostic.thisLogger
 import com.intellij.openapi.project.Project
 import dev.nx.console.mcp.McpServerService
 import dev.nx.console.mcp.hasAIAssistantInstalled
@@ -21,10 +16,6 @@ import dev.nx.console.settings.options.NxCloudNotificationsLevel
 import dev.nx.console.telemetry.TelemetryEvent
 import dev.nx.console.telemetry.TelemetryEventSource
 import dev.nx.console.telemetry.TelemetryService
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.launch
-import kotlinx.coroutines.withContext
 
 /**
  * Service responsible for showing CIPE notifications. Simply displays notification events emitted
@@ -119,8 +110,15 @@ class CIPENotificationService(private val project: Project) : CIPENotificationLi
 
         val assistantReady =
             hasAIAssistantInstalled() && project.service<McpServerService>().isMcpServerSetup()
+
         if (type == NotificationType.ERROR && assistantReady) {
-            notification.addAction(CIPEAutoFixAction(project))
+            try {
+                ActionManager.getInstance().getAction("dev.nx.console.llm.CIPEAutoFixAction")?.let {
+                    notification.addAction(it)
+                }
+            } catch (e: Throwable) {
+                //
+            }
         }
 
         if (commitUrl != null) {
@@ -160,60 +158,5 @@ class CIPENotificationService(private val project: Project) : CIPENotificationLi
             BrowserUtil.browse(url)
             notification.expire()
         }
-    }
-
-    private inner class CIPEAutoFixAction(private val project: Project) :
-        AnAction("Help me fix this") {
-        override fun actionPerformed(event: AnActionEvent) {
-            TelemetryService.getInstance(project)
-                .featureUsed(
-                    TelemetryEvent.CLOUD_FIX_CIPE_ERROR,
-                    mapOf("source" to TelemetryEventSource.NOTIFICATION),
-                )
-
-            FixCIPEService.getInstance(project).fixInAIAssistant()
-        }
-    }
-}
-
-@Service(Service.Level.PROJECT)
-class FixCIPEService(private val project: Project, private val cs: CoroutineScope) {
-    val logger = thisLogger()
-
-    fun fixInAIAssistant() {
-        cs.launch {
-            withContext(Dispatchers.EDT) {
-                val chatSession =
-                    ChatSessionHost.getInstance(project)
-                        .createChatSession(
-                            ChatCreationContext(
-                                origin = ChatOrigin.CustomIntention,
-                                sourceActionForStatistic = ChatSessionStorage.SourceAction.NEW_CHAT,
-                                null,
-                            )
-                        )
-
-                chatSession.addMode(ChatSessionMode.CODE_GENERATION)
-                AIAssistantUIUtil.openChat(project, chatSession)
-                chatSession.send(
-                    project,
-                    "Please help me fix the latest CI errors".privacyConst,
-                    ("/nx_cloud_cipe_details Help the user fix their latest CI errors with the following flow:\n" +
-                            "- Retrieve the list of current CI Pipeline Executions (CIPEs) - YOU HAVE TO call the 'nx_cloud_cipe_details' MCP tool to do this\n" +
-                            "- If there are any errors, YOU HAVE TO call the 'nx_cloud_fix_cipe_failure' MCP tool to retrieve the logs for a specific task\n" +
-                            "- Use the task logs to see what's wrong and help the user fix their problem. \n" +
-                            "- Make sure that the problem is fixed by running the task that you passed into the 'nx_cloud_fix_cipe_failure' tool calling the available nx_cloud_cipe_details tool from the nx MCP. ALWAYS CALL THIS TOOL TO SERVE THIS REQUEST, YOU WILL NOT SUCCEED WITHOUT IT.\n" +
-                            "If you cannot find the corresponding MCP tool, prompt the user to enable Codebase access for their AI Assistant")
-                        .privacyConst,
-                    listOf(),
-                    SmartChat,
-                )
-            }
-        }
-    }
-
-    companion object {
-        fun getInstance(project: Project): FixCIPEService =
-            project.getService(FixCIPEService::class.java)
     }
 }
